Update security and docker-compose.

+++ src/main/java/stu/napls/napurity/repository/UserRepository.java
@@ -3,6 +3,5 @@ package stu.napls.napurity.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import stu.napls.napurity.model.User;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
-
+public interface UserRepository extends JpaRepository<User, Long> {
 }
